Move embedded test resources to fixtures
